Split JsonArtifactReader (1959 LOC) into a main class plus three helpers
Extract three package-private helper classes from the monolithic
JsonArtifactReader so the file can be navigated:

- JsonNodeReaders (27 methods): primitive node accessors -- readString,
  readUri, readVersion, readOffsetDateTime, read{Required,Optional,...}*,
  read{String,Uri}Array, read*Map, plus tiny helpers like
  hasJsonLdContextField and getFieldNode. All stateless.

- JsonValueConstraintsReader (18 methods): the value-constraints parsing
  layer -- the 5 source types (ontology / class / branch / value-set /
  literal), their default values and actions, and the enum-readers for
  ValueType / ValueConstraintsActionType / XsdNumericDatatype /
  XsdTemporalDatatype.

- JsonArtifactShapeChecks (7 methods + 2 constants): JSON Schema/JSON-LD
  shape validation -- check*JsonLdType for template/element/field,
  checkArtifactJsonSchemaType, checkArtifactJsonSchemaSchemaUri,
  checkSchemaArtifactModelVersion. Constants `MODEL_VERSION` and
  `JSON_SCHEMA_SCHEMA_URI` move here from instance fields on the reader.

All extracted methods become `public static` (package-private use only;
class itself is package-private and final with a private constructor).
JsonArtifactReader pulls them in via `import static`, so no call-site
edits were needed -- bare `readString(...)` still resolves.

JsonArtifactReader.java: 1959 -> 999 LOC (-49%).

mvn clean test: 270 tests, 0 failures, 3 pre-existing skips.
mvn spotless:check: clean.
